exec: Fix potential race condition in boost task checker
A race condition where the task's comm is updated while it is being
compared may occur because we check the comm directly. Fix this by using
the get_task_comm helper which copies the comm for comparison.

Additionally, move and rename the checker function to reflect the change
in arguments.

Change-Id: I092657e2f58a479ff9e3ad70e5369840f8239854
Signed-off-by: Danny Lin <danny@kdrag0n.dev>